% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{combine_incidence_fits}
\alias{combine_incidence_fits}
\title{Combine the posteriors of multiple stan fit}
\usage{
combine_incidence_fits(incidence_list)
}
\arguments{
\item{incidence_list}{A list of incidence model fits converted to be
\code{stanfit} objects from th \code{rstan} packag}
}
\value{
A \code{stanfit} model with combined posterior estimates
}
\description{
Combine the posteriors of multiple stan fits into
a single \code{rstan} \code{stanfit} object. This is likely useful when using
\code{incidence_lapply()} to fit models using multiple sampled probability
of detection curves. A simple wrapper around \code{rstan::sflist2stanfit()}
}
\details{
Progression information is displayed using the
\code{progresr} package. See the package documentation for options.
}
\seealso{
Other incidence: 
\code{\link{incidence_lapply}()},
\code{\link{incidence}()}
}
\concept{incidence}
