% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-tools.R
\name{extract_dated_parameter}
\alias{extract_dated_parameter}
\title{Extract samples for a parameter with a date dimension from a Stan model}
\usage{
extract_dated_parameter(fit, var, start_date)
}
\arguments{
\item{fit}{An \code{rstanfit} object produced by \code{incidence()}}

\item{var}{Character string, variable to plot.}

\item{start_date}{A date, used to index the time plot.}
}
\value{
A data frame containing the parameter name, date, sample id and
sample value
}
\description{
Extracts a single parameter with a date dimension from a list
of \code{stan} output and returns it as a \code{data.table}.
}
