% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior-tools.R
\name{extract_static_parameter}
\alias{extract_static_parameter}
\title{Extract samples from a parameter with a single dimension}
\usage{
extract_static_parameter(param, samples)
}
\value{
A data frame containing the parameter name, sample id and sample
value.
}
\description{
Extract samples from a parameter with a single dimension
}
