% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{i2p_model}
\alias{i2p_model}
\title{Load and compile the model}
\usage{
i2p_model(
  model = "stan/inc2prev.stan",
  include,
  compile = TRUE,
  threads = FALSE,
  verbose = TRUE,
  optimise = TRUE,
  ...
)
}
\arguments{
\item{model}{A character string indicating the path to the model.
If not supplied the package default model is used.}

\item{include}{A character string specifying the path to any stan
files to include in the model. If missing the package default is used.}

\item{compile}{Logical, defaults to \code{TRUE}. Should the model
be loaded and compiled using \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}

\item{threads}{Logical, defaults to \code{FALSE}. Should the model compile with
support for multi-thread support in chain. Note that this requires the use of
the \code{threads_per_chain} argument when model fitting using \code{\link[=enw_sample]{enw_sample()}},
and \code{\link[=epinowcast]{epinowcast()}}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose
messages be shown.}

\item{...}{Additional arguments passed to \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}
}
\value{
A \code{cmdstanr} model.
}
\description{
Load and compile the model
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mod <- i2p_model()
\dontshow{\}) # examplesIf}
}
\seealso{
Other model: 
\code{\link{i2p_sample}()}
}
\concept{model}
