% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{i2p_sample}
\alias{i2p_sample}
\title{Fit a CmdStan model using NUTS}
\usage{
i2p_sample(data, model = i2p_model(), diagnostics = TRUE, ...)
}
\arguments{
\item{data}{A list of data as produced by \code{\link[=enw_as_data_list]{enw_as_data_list()}}.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=enw_model]{enw_model()}}.}

\item{diagnostics}{Logical, defaults to \code{TRUE}. Should fitting diagnostics
be returned as a \code{data.frame}.}

\item{...}{Additional parameters passed to the \code{sample} method of \code{cmdstanr}.}
}
\value{
A \code{data.frame} containing the \code{cmdstanr} fit, the input data, the
fitting arguments, and optionally summary diagnostics.
}
\description{
Fit a CmdStan model using NUTS
}
\seealso{
Other model: 
\code{\link{i2p_model}()}
}
\concept{model}
