% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc2prev-package.R
\docType{package}
\name{inc2prev-package}
\alias{inc2prev}
\alias{inc2prev-package}
\title{inc2prev: Estimate incidence from Prevalence Estimates}
\description{
Estimate incidence from prevalence estimates, currently demonstrated using the Office for National Statistics COVID-19 prevalence estimates.
}
\author{
\strong{Maintainer}: Sam Abbott \email{sam.abbott@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

Authors:
\itemize{
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID})
}

}
\keyword{internal}
