% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence_lapply}
\alias{incidence_lapply}
\title{Fit multiple incidence models efficiently}
\usage{
incidence_lapply(prev_list, ...)
}
\arguments{
\item{prev_list}{A list of lists of data as produced by \code{i2p_data()}.}

\item{...}{Additional arguments passed to \code{incidence()} and/or
\code{future_lapply()}.}
}
\value{
A fit stan model as produced by the option passed to
\code{fit_fn}.
}
\description{
Fits multiple \code{incidence()} models in an efficient,
optionally in parallel, framework with user definable progression
settings.
}
\details{
Progression information is displayed using the
\code{progressr} package. See the package documentation for options.
}
\seealso{
Other incidence: 
\code{\link{combine_incidence_fits}()},
\code{\link{incidence}()}
}
\concept{incidence}
