% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse logit function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{a logit values or vector of logit values}
}
\value{
the probability calculated as inverse logit
}
\description{
The inverse logit function
}
\examples{
inv_logit(c(-10, 1, 0, 100))
}
\seealso{
Other utils: 
\code{\link{expose_stan_fns}()},
\code{\link{logit}()}
}
\concept{utils}
