% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{a probability or vector of probabilities}
}
\value{
the logit of the probability
}
\description{
The logit function
}
\examples{
logit(c(0.01, 0.1, 0.5, 0.9, 1))
}
\seealso{
Other utils: 
\code{\link{expose_stan_fns}()},
\code{\link{inv_logit}()}
}
\concept{utils}
