% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.inc2prev}
\alias{plot.inc2prev}
\alias{plot}
\title{Plot method for inc2prev}
\usage{
\method{plot}{inc2prev}(type = "trace", ...)
}
\arguments{
\item{type}{A character vector indicating the plot method to use. Supported
options are "cri" (plot credible intervals), "trace" (plot posterior
samples), and "obs" (plot observations against the posterior).}

\item{...}{Pass additional arguments to underlying plot functions}

\item{x}{A model fit using the \code{incidence()} method.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot method for models fit using the \code{inc2prev} package.
}
