% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cri}
\alias{plot_cri}
\title{Plot posterior credible intervals over time}
\usage{
plot_cri(fit, var, start_date)
}
\arguments{
\item{fit}{An \code{rstanfit} object produced by \code{incidence()}}

\item{var}{Character string, variable to plot.}

\item{start_date}{A date, used to index the time plot.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot posterior credible intervals over time
}
