% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_obs}
\alias{plot_obs}
\title{Plot observations against posterior samples over time}
\usage{
plot_obs(fit, prev, samples = 100, date_start, alpha = 0.05, data_source)
}
\arguments{
\item{fit}{An \code{rstanfit} object produced by \code{incidence()}}

\item{samples}{Integer, defaults to 100. The number of posterior
samples to extract.}

\item{alpha}{Numeric, defaults to 0.05. The alpha value to use in
the\code{ggplot2}  plot for the individual posterior traces.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot observations against posterior samples over time
}
