% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot posterior samples over time}
\usage{
plot_trace(fit, var, start_date, samples = 100, alpha = 0.05)
}
\arguments{
\item{samples}{Integer, defaults to 100. The number of posterior
samples to extract.}

\item{alpha}{Numeric, defaults to 0.05. The alpha value to use in
the\code{ggplot2}  plot for the individual posterior traces.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot posterior samples over time
}
