% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot summarised trend over time in posteriors}
\usage{
plot_trend(fit, var, start_date)
}
\value{
A \code{ggplot2} object
}
\description{
Plot summarised trend over time in posteriors
}
